sanguo/requirements.txt
scrapy_splash==0.7.2
Scrapy==1.8.0
mysql_connector_repackaged==0.3.1